# [aggregator] Move some struct members around using fieldalignment tool