--- a/src/aggregator/aggregation/gauge.go
+++ b/src/aggregator/aggregation/gauge.go
@@ -30,6 +30,7 @@ import (
 // Gauge aggregates gauge values.
 type Gauge struct {
 	Options
+
 	lastAt     time.Time
 	annotation []byte
 	sum        float64