--- a/src/metrics/metric/aggregated/types.go
+++ b/src/metrics/metric/aggregated/types.go
@@ -40,11 +40,11 @@ var (
 
 // Metric is a metric, which is essentially a named value at certain time.
 type Metric struct {
-	Type       metric.Type
 	ID         id.RawID
+	Annotation []byte
+	Type       metric.Type
 	TimeNanos  int64
 	Value      float64
-	Annotation []byte
 }
 
 // ToProto converts the metric to a protobuf message in place.
@@ -84,9 +84,9 @@ func (m Metric) String() string {
 // ChunkedMetric is a metric with a chunked ID.
 type ChunkedMetric struct {
 	id.ChunkedID
+	Annotation []byte
 	TimeNanos  int64
 	Value      float64
-	Annotation []byte
 }
 
 // RawMetric is a metric in its raw form (e.g., encoded bytes associated with
@@ -147,12 +147,12 @@ type ChunkedMetricWithStoragePolicy struct {
 
 // ForwardedMetric is a forwarded metric.
 type ForwardedMetric struct {
-	Type       metric.Type
 	ID         id.RawID
-	TimeNanos  int64
 	Values     []float64
 	PrevValues []float64
 	Annotation []byte
+	Type       metric.Type
+	TimeNanos  int64
 	Version    uint32
 }
 
@@ -247,8 +247,8 @@ func (tm *TimedMetricWithMetadata) FromProto(pb *metricpb.TimedMetricWithMetadat
 
 // TimedMetricWithMetadatas is a timed metric with staged metadatas.
 type TimedMetricWithMetadatas struct {
-	Metric
 	metadata.StagedMetadatas
+	Metric
 }
 
 // ToProto converts the timed metric with metadata to a protobuf message in place.