--- a/src/metrics/metric/id/m3/id.go
+++ b/src/metrics/metric/id/m3/id.go
@@ -103,8 +103,8 @@ func IsRollupID(name []byte, tags []byte, iterPool id.SortedTagIteratorPool) boo
 
 // TODO(xichen): pool the mids.
 type metricID struct {
-	id       []byte
 	iterPool id.SortedTagIteratorPool
+	id       []byte
 }
 
 // NewID creates a new m3 metric id.
@@ -150,12 +150,12 @@ func NameAndTags(id []byte) ([]byte, []byte, error) {
 }
 
 type sortedTagIterator struct {
-	sortedTagPairs []byte
-	idx            int
-	tagName        []byte
-	tagValue       []byte
 	err            error
 	pool           id.SortedTagIteratorPool
+	sortedTagPairs []byte
+	tagName        []byte
+	tagValue       []byte
+	idx            int
 }
 
 // NewSortedTagIterator creates a new sorted tag iterator.