Went through a couple of packages with `fieldalignment -fix` that are commonly used. Unlike `maligned`, it also moves pointers to the front of the struct to reduce amount of bytes GC has to scan.
Microbenchmarks (that exist) are really boring and don't show anything, except for one that churns through a lot of data and does actually GC:
```
name                                                              old time/op    new time/op    delta
TimerAddBatch/100k_samples_1000_batch_size-12                       1.92ms ± 7%    1.65ms ± 4%  -13.83%  (p=0.008 n=5+5)
TimerAddBatch/100k_samples_1000_batch_size_0..1,000,000_range-12    1.92ms ± 3%    1.67ms ± 7%  -12.88%  (p=0.008 n=5+5)
TimerAddBatch/100k_samples_1000_batch_size_0..100,000_range-12      1.93ms ± 6%    1.83ms ± 1%   -5.10%  (p=0.016 n=5+5)
TimerAddBatch/10k_samples_1000_batch_size-12                         315µs ± 8%     291µs ± 2%   -7.61%  (p=0.008 n=5+5)
TimerAddBatch/10k_samples_100_batch_size-12                          315µs ± 2%     300µs ± 1%   -4.99%  (p=0.008 n=5+5)
TimerAddBatch/10k_samples_100_batch_size_with_negative_values-12     143µs ± 2%     142µs ± 2%     ~     (p=1.000 n=5+5)
TimerAddBatch/1k_samples_100_batch_size-12                          12.5µs ± 1%    12.1µs ± 2%   -3.40%  (p=0.008 n=5+5)

name                                                              old speed      new speed      delta
TimerAddBatch/100k_samples_1000_batch_size-12                      418MB/s ± 6%   484MB/s ± 4%  +15.94%  (p=0.008 n=5+5)
TimerAddBatch/100k_samples_1000_batch_size_0..1,000,000_range-12   418MB/s ± 3%   480MB/s ± 7%  +14.90%  (p=0.008 n=5+5)
TimerAddBatch/100k_samples_1000_batch_size_0..100,000_range-12     416MB/s ± 6%   438MB/s ± 1%   +5.25%  (p=0.016 n=5+5)
TimerAddBatch/10k_samples_1000_batch_size-12                       255MB/s ± 7%   275MB/s ± 2%   +8.09%  (p=0.008 n=5+5)
TimerAddBatch/10k_samples_100_batch_size-12                        254MB/s ± 2%   267MB/s ± 1%   +5.22%  (p=0.008 n=5+5)
TimerAddBatch/10k_samples_100_batch_size_with_negative_values-12   561MB/s ± 2%   563MB/s ± 2%     ~     (p=1.000 n=5+5)
TimerAddBatch/1k_samples_100_batch_size-12                         639MB/s ± 1%   662MB/s ± 2%   +3.54%  (p=0.008 n=5+5)

name                                                              old alloc/op   new alloc/op   delta
TimerAddBatch/100k_samples_1000_batch_size-12                        527kB ± 0%     527kB ± 0%     ~     (p=0.222 n=5+5)
TimerAddBatch/100k_samples_1000_batch_size_0..1,000,000_range-12     527kB ± 0%     527kB ± 0%     ~     (p=0.421 n=5+5)
TimerAddBatch/100k_samples_1000_batch_size_0..100,000_range-12       527kB ± 0%     527kB ± 0%     ~     (p=1.000 n=5+5)
TimerAddBatch/10k_samples_1000_batch_size-12                         531kB ± 0%     531kB ± 0%     ~     (p=0.548 n=5+5)
TimerAddBatch/10k_samples_100_batch_size-12                          530kB ± 0%     530kB ± 0%     ~     (p=0.548 n=5+5)
TimerAddBatch/10k_samples_100_batch_size_with_negative_values-12      197B ± 1%      200B ± 2%     ~     (p=0.079 n=5+5)
TimerAddBatch/1k_samples_100_batch_size-12                           96.0B ± 0%     96.0B ± 0%     ~     (all equal)
```

Report output:
```
src/aggregator/aggregation/counter.go:31:14: struct with 112 pointer bytes could be 72
src/aggregator/aggregation/gauge.go:31:12: struct with 120 pointer bytes could be 72
src/aggregator/aggregation/options.go:35:14: struct with 40 pointer bytes could be 32
src/aggregator/aggregation/timer.go:31:12: struct with 64 pointer bytes could be 40

src/aggregator/bitset/bitset.go:30:13: struct with 24 pointer bytes could be 8

src/aggregator/client/payload.go:42:21: struct with 128 pointer bytes could be 120
src/aggregator/client/payload.go:67:19: struct with 600 pointer bytes could be 592
src/aggregator/client/queue.go:123:12: struct with 192 pointer bytes could be 144
src/aggregator/client/writer.go:64:13: struct with 160 pointer bytes could be 120
src/aggregator/client/writer.go:416:23: struct with 32 pointer bytes could be 24

src/aggregator/aggregation/quantile/cm/options.go:39:14: struct with 32 pointer bytes could be 8
src/aggregator/aggregation/quantile/cm/sample.go:24:13: struct with 40 pointer bytes could be 16
src/aggregator/aggregation/quantile/cm/stream.go:41:13: struct with 248 pointer bytes could be 192

src/metrics/pipeline/type.go:195:15: struct with 48 pointer bytes could be 32
src/metrics/pipeline/type.go:407:22: struct with 32 pointer bytes could be 24
src/metrics/pipeline/type.go:428:14: struct with 72 pointer bytes could be 48
src/metrics/pipeline/applied/type.go:84:14: struct with 24 pointer bytes could be 8
src/metrics/metadata/metadata.go:77:23: struct with 96 pointer bytes could be 80
src/metrics/metadata/metadata.go:394:22: struct of size 80 could be 72
src/metrics/metadata/metadata.go:651:31: struct with 16 pointer bytes could be 8
src/metrics/transformation/type.go:223:9: struct with 56 pointer bytes could be 48
src/metrics/encoding/protobuf/aggregated_encoder.go:39:24: struct with 128 pointer bytes could be 96
src/metrics/encoding/protobuf/buffer.go:33:13: struct with 32 pointer bytes could be 16
src/metrics/encoding/protobuf/unaggregated_encoder.go:61:26: struct with 848 pointer bytes could be 776
src/metrics/encoding/protobuf/unaggregated_iterator.go:34:27: struct with 928 pointer bytes could be 848
src/metrics/metric/aggregated/types.go:42:13: struct with 56 pointer bytes could be 32
src/metrics/metric/aggregated/types.go:85:20: struct with 96 pointer bytes could be 80
src/metrics/metric/aggregated/types.go:149:22: struct with 96 pointer bytes could be 80
src/metrics/metric/id/m3/id.go:105:15: struct with 40 pointer bytes could be 24
src/metrics/metric/id/m3/id.go:152:24: struct with 112 pointer bytes could be 88
src/metrics/metric/unaggregated/types.go:43:14: struct with 40 pointer bytes could be 32
src/metrics/metric/unaggregated/types.go:112:12: struct with 40 pointer bytes could be 32
src/metrics/metric/unaggregated/types.go:147:30: struct with 72 pointer bytes could be 64
src/metrics/metric/unaggregated/types.go:153:33: struct with 88 pointer bytes could be 80
src/metrics/metric/unaggregated/types.go:159:28: struct with 72 pointer bytes could be 64
src/metrics/metric/unaggregated/types.go:165:27: struct with 72 pointer bytes could be 64
src/metrics/metric/unaggregated/types.go:192:30: struct with 88 pointer bytes could be 80
src/metrics/metric/unaggregated/types.go:219:25: struct with 72 pointer bytes could be 64
src/metrics/metric/unaggregated/types.go:250:18: struct with 96 pointer bytes could be 72
```